Refactoring to a separate commands file
